refactor: rename interfaces.ts as types.ts and move file to root folder
