+++ src/store/thunks/updateProfileLanguagesThunk.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { resumes } from '../../api/student/getStudentDetailWithId';
-import { TLanguage } from '../../interfaces/interfaces';
+import { TLanguage } from '../../../types';
 
 type DataFetch = {
   name: string,
